moved redeem stake out of msg_server
